requirements.txt
# web-api-server
fastapi
uvicorn

# mecab
mecab-python3
unidic
